+++ am_list/src/python.rs
@@ -1,6 +1,7 @@
 mod queries;
 
-use crate::{FunctionInfo, ListAmFunctions, Result};
+use crate::{FunctionInfo, InstrumentFile, ListAmFunctions, Result};
+use log::debug;
 use queries::{AllFunctionsQuery, AmImportQuery, AmQuery};
 use rayon::prelude::*;
 use std::{
@@ -33,6 +34,41 @@ impl Impl {
                 .extension()
                 .map_or(false, |ext| ext == "py" || ext == "py3")
     }
+
+    fn list_files(
+        project_root: &Path,
+        exclude_patterns: Option<&ignore::gitignore::Gitignore>,
+    ) -> Vec<String> {
+        const PREALLOCATED_ELEMS: usize = 100;
+        let walker = WalkDir::new(project_root).into_iter();
+        let mut project_files = Vec::with_capacity(PREALLOCATED_ELEMS);
+        project_files.extend(walker.filter_entry(Self::is_valid).filter_map(|entry| {
+            let entry = entry.ok()?;
+
+            if let Some(pattern) = exclude_patterns {
+                let ignore_match =
+                    pattern.matched_path_or_any_parents(entry.path(), entry.file_type().is_dir());
+                if matches!(ignore_match, ignore::Match::Ignore(_)) {
+                    debug!(
+                        "The exclusion pattern got a match on {}: {:?}",
+                        entry.path().display(),
+                        ignore_match
+                    );
+                    return None;
+                }
+            }
+
+            Some(
+                entry
+                    .path()
+                    .to_str()
+                    .map(ToString::to_string)
+                    .unwrap_or_default(),
+            )
+        }));
+
+        project_files
+    }
 }
 
 impl ListAmFunctions for Impl {
@@ -43,21 +79,9 @@ impl ListAmFunctions for Impl {
             .file_name()
             .map(|s| s.to_str().unwrap_or_default())
             .unwrap_or("");
+        let project_files = Self::list_files(project_root, None);
 
-        let walker = WalkDir::new(project_root).into_iter();
-        let mut source_mod_pairs = Vec::with_capacity(PREALLOCATED_ELEMS);
-        source_mod_pairs.extend(walker.filter_entry(Self::is_valid).filter_map(|entry| {
-            let entry = entry.ok()?;
-            Some(
-                entry
-                    .path()
-                    .to_str()
-                    .map(ToString::to_string)
-                    .unwrap_or_default(),
-            )
-        }));
-
-        list.par_extend(source_mod_pairs.par_iter().filter_map(move |path| {
+        list.par_extend(project_files.par_iter().filter_map(move |path| {
             let relative_module_name = Path::new(path)
                 .strip_prefix(project_root)
                 .ok()?
@@ -94,20 +118,9 @@ impl ListAmFunctions for Impl {
             .map(|s| s.to_str().unwrap_or_default())
             .unwrap_or("");
 
-        let walker = WalkDir::new(project_root).into_iter();
-        let mut source_mod_pairs = Vec::with_capacity(PREALLOCATED_ELEMS);
-        source_mod_pairs.extend(walker.filter_entry(Self::is_valid).filter_map(|entry| {
-            let entry = entry.ok()?;
-            Some(
-                entry
-                    .path()
-                    .to_str()
-                    .map(ToString::to_string)
-                    .unwrap_or_default(),
-            )
-        }));
+        let project_files = Self::list_files(project_root, None);
 
-        list.par_extend(source_mod_pairs.par_iter().filter_map(move |path| {
+        list.par_extend(project_files.par_iter().filter_map(move |path| {
             let relative_module_name = Path::new(path)
                 .strip_prefix(project_root)
                 .ok()?
@@ -132,6 +145,98 @@ impl ListAmFunctions for Impl {
         let mut result = Vec::with_capacity(PREALLOCATED_ELEMS);
         result.extend(list.into_iter().flatten());
         Ok(result)
+    }
+
+    fn list_autometrics_functions_in_single_file(
+        &mut self,
+        source_code: &str,
+    ) -> Result<Vec<FunctionInfo>> {
+        let import_query = AmImportQuery::try_new()?;
+        let decorator_name = import_query.get_decorator_name(source_code).ok();
+        if decorator_name.is_none() {
+            return Ok(Vec::new());
+        }
+        let query = AmQuery::try_new(decorator_name.as_ref().unwrap())?;
+        query.list_function_names("<single file>", source_code, "")
+    }
+
+    fn list_all_function_definitions_in_single_file(
+        &mut self,
+        source_code: &str,
+    ) -> Result<Vec<FunctionInfo>> {
+        let query = AllFunctionsQuery::try_new()?;
+        query.list_function_names("<single file>", source_code, "")
+    }
+}
+
+impl InstrumentFile for Impl {
+    fn instrument_source_code(&mut self, source: &str) -> Result<String> {
+        const DEF_LEN: usize = "def ".len();
+
+        let mut locations = self.list_all_functions_in_single_file(source)?;
+        locations.sort_by_key(|info| {
+            info.definition
+                .as_ref()
+                .map(|def| def.range.start.line)
+                .unwrap_or_default()
+        });
+
+        let has_am_directive = source
+            .lines()
+            .any(|line| line.contains("from autometrics import autometrics"));
+
+        let mut new_code = crop::Rope::from(source);
+        // Keeping track of inserted lines to update the byte offset to insert code to,
+        // only works if the locations list is sorted from top to bottom
+        let mut inserted_lines = 0;
+
+        if !has_am_directive {
+            new_code.insert(0, "from autometrics import autometrics\n");
+            inserted_lines += 1;
+        }
+
+        for function_info in locations {
+            if function_info.definition.is_none() || function_info.instrumentation.is_some() {
+                continue;
+            }
+
+            let def_line = function_info.definition.as_ref().unwrap().range.start.line;
+            let def_col = function_info
+                .definition
+                .unwrap()
+                .range
+                .start
+                .column
+                .saturating_sub(DEF_LEN);
+            let byte_offset = new_code.byte_of_line(inserted_lines + def_line);
+            new_code.insert(
+                byte_offset,
+                format!("{}@autometrics\n", " ".repeat(def_col)),
+            );
+            inserted_lines += 1;
+        }
+
+        Ok(new_code.to_string())
+    }
+
+    fn instrument_project(
+        &mut self,
+        project_root: &Path,
+        exclude_patterns: Option<&ignore::gitignore::Gitignore>,
+    ) -> Result<()> {
+        let sources_modules = Self::list_files(project_root, exclude_patterns);
+
+        for path in sources_modules {
+            if std::fs::metadata(&path)?.is_dir() {
+                continue;
+            }
+            debug!("Instrumenting {path}");
+            let old_source = read_to_string(&path)?;
+            let new_source = self.instrument_source_code(&old_source)?;
+            std::fs::write(path, new_source)?;
+        }
+
+        Ok(())
     }
 }